requirements.txt
streamlit==1.44.1
numpy==2.0.2
matplotlib==3.9.4
sqlalchemy==2.0.40
pydantic==2.11.3
pillow==11.1.0
opencv-python==4.11.0.86
absl-py==2.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf>=3.20,<6
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
streamlit-image-comparison==0.0.4
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.102
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0


trycourier~=4.4.0
alembic~=1.15.2
passlib~=1.7.4